fix: suppress browser launch in noui builds (#812)
