+++ .github/workflows/ci.yml
@@ -140,9 +140,9 @@ jobs:
         with:
           go-version: ${{ matrix.go-version }}
       - name: build go
-        run: go build ./cmd/app/main.go
+        run: go build ./cmd/app
       - name: build noui (headless)
-        run: CGO_ENABLED=0 go build -tags=noui -o console-headless ./cmd/app/main.go
+        run: CGO_ENABLED=0 go build -tags=noui -o console-headless ./cmd/app
       - name: Install Test Converter and run tests
         run: |
           export GOPATH="$HOME/go/"
